backend

+++ BackEnd_SebastianChaile/src/main/java/com/portfolio/szk/Entity/Persona.java
@@ -22,7 +22,7 @@ import lombok.Setter;
 public class Persona {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private int id;
     
     @NotNull
     @Size(min=1,max=50,message="no cumple con la longitud")
@@ -36,6 +36,6 @@ public class Persona {
     @Size(min=1,max=50,message="no cumple con la longitud")
     private String img;
     
-    
-    
+
+
 }
